Added Appwrite and Hilt dependency. Started de registration part.
