+++ app/src/main/java/com/dersarco/petpalplaces/di/AppModule.kt
@@ -2,7 +2,8 @@ package com.dersarco.petpalplaces.di
 
 import android.content.Context
 import com.dersarco.petpalplaces.data.remote.AppWrite
-import com.dersarco.petpalplaces.domain.usecase.RegisterUserUseCase
+import com.dersarco.petpalplaces.domain.usecase.RegisterFacebookUserUseCase
+import com.dersarco.petpalplaces.domain.usecase.RegisterGoogleUserUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -23,6 +24,10 @@ object AppModule {
 
     @Provides
     @Singleton
-    fun provideRegisterUserUseCase(client: Client) = RegisterUserUseCase(client)
+    fun provideRegisterFacebookUserUseCase(client: Client) = RegisterFacebookUserUseCase(client)
+
+    @Provides
+    @Singleton
+    fun provideRegisterGoogleUserUseCase(client: Client) = RegisterGoogleUserUseCase(client)
 
 }
