+++ app/src/main/java/com/dersarco/petpalplaces/domain/utils/AppwriteConst.kt
@@ -1,0 +1,9 @@
+package com.dersarco.petpalplaces.domain.utils
+
+object AppwriteConst {
+    const val SUCCESS =
+        "appwrite-callback-6464eeabe592f858144b://cloud.appwrite.io/auth/oauth2/success"
+    const val FAILURE =
+        "appwrite-callback-6464eeabe592f858144b://cloud.appwrite.io/auth/oauth2/failure"
+
+}
