+++ app/src/main/java/com/dersarco/petpalplaces/ui/screens/register/RegisterScreen.kt
@@ -42,7 +42,7 @@ fun RegisterScreen(
                     LogoImage(logo = R.drawable.logo)
                 }
                 RegisterForm(padding) {
-                    viewModel.registerUser(context as ComponentActivity)
+                    viewModel.registerGoogleUser(context as ComponentActivity)
                 }
             }
         }
