+++ app/src/main/java/com/dersarco/petpalplaces/ui/screens/register/RegisterViewModel.kt
@@ -3,7 +3,8 @@ package com.dersarco.petpalplaces.ui.screens.register
 import androidx.activity.ComponentActivity
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.dersarco.petpalplaces.domain.usecase.RegisterUserUseCase
+import com.dersarco.petpalplaces.domain.usecase.RegisterFacebookUserUseCase
+import com.dersarco.petpalplaces.domain.usecase.RegisterGoogleUserUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -11,12 +12,19 @@ import javax.inject.Inject
 
 @HiltViewModel
 class RegisterViewModel @Inject constructor(
-    private val registerUserUseCase: RegisterUserUseCase
+    private val registerFacebookUserUseCase: RegisterFacebookUserUseCase,
+    private val registerGoogleUserUseCase: RegisterGoogleUserUseCase
 ) : ViewModel() {
 
-    fun registerUser(activity: ComponentActivity) {
+    fun registerFacebookUser(activity: ComponentActivity) {
         viewModelScope.launch(Dispatchers.Main) {
-            registerUserUseCase(activity)
+            registerFacebookUserUseCase(activity)
+        }
+    }
+
+    fun registerGoogleUser(activity: ComponentActivity) {
+        viewModelScope.launch(Dispatchers.Main) {
+            registerGoogleUserUseCase(activity)
         }
     }
 }
