+++ build.gradle
@@ -2,6 +2,6 @@
 plugins {
     id 'com.android.application' version '8.0.0' apply false
     id 'com.android.library' version '8.0.0' apply false
-    id 'org.jetbrains.kotlin.android' version '1.8.10' apply false
+    id 'org.jetbrains.kotlin.android' version '1.9.0-Beta' apply false
     id 'com.google.dagger.hilt.android' version '2.44' apply false
 }
